refactor: lift reviews up
